More clean up, no more warnings, happy to go to bed - the triangles are rendering and all seems good - will do code review tomorrow and take my time :)
